+++ dtlpymetrics/temp/check_video_scoring.py
@@ -3,8 +3,7 @@ import os
 import logging
 
 import dtlpy as dl
-from dtlpymetrics.scoring import calculate_task_score
-from dtlpymetrics.dtlpy_scores import ScoreType, Scores, Score
+from dtlpymetrics.scoring import calculate_task_score, create_task_item_score
 
 logger = logging.getLogger()
 
@@ -25,5 +24,13 @@ qualification_task_scores = calculate_task_score(task=qualification_task)
 # dl_scores = Scores(client_api=dl.client_api)
 # dl_scores.get(score_id='64c8f0e4a3f0796c68c35532')
 
-input('Press enter to continue...')
 print()
+
+
+# vid_gt = dl.items.get(item_id='64be6ae759c68826d2488345')
+# vid_test = dl.items.get(item_id='64c8f0f9b118f7360a1a071d')
+#
+# qualification_task = dl.tasks.get(task_id='64c8f0e4a3f0796c68c35532')
+#
+# vid_item_scores = create_task_item_score(item=vid_gt,
+#                                          task=qualification_task)
